fix: move ApiTipHistorySource to tipHistorySource.ts and add defensive checks
