+++ frontend/src/__tests__/tipHistorySource.test.ts
@@ -1,6 +1,5 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { FixtureTipHistorySource } from '../services/tipHistorySource';
-import { ApiTipHistorySource } from '../services/tipService';
+import { FixtureTipHistorySource, ApiTipHistorySource } from '../services/tipHistorySource';
 import { mockTipHistoryData } from '../fixtures/tipHistory.fixtures';
 import type { TipFiltersState } from '../components/tip-history';
